create home,regster view + modify loading screen + create a component loadingScreen
