+++ index.html
@@ -6,15 +6,14 @@
   <link rel="icon" type="image/svg+xml" href="/vite.svg" />
   <link rel="stylesheet" href="sass/style.css">
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Count4Me</title>
+  <title>Countify</title>
 </head>
 
 <body>
   <div id="root">
-    <div class="container">
+    <!-- <div class="container">
       <div class="🍺"></div>
-      <h1 class="titleApp">LOADING...</h1>
-    </div>
+    </div> -->
   </div>
   <script type="module" src="/src/main.jsx"></script>
   <script src="https://kit.fontawesome.com/9af8e35030.js" crossorigin="anonymous"></script>
