+++ package.json
@@ -15,6 +15,8 @@
     "firebase": "^9.19.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
+    "sass": "^1.62.0",
     "three": "^0.150.1"
   },
   "devDependencies": {
