+++ public/sass/style.scss
@@ -5,13 +5,16 @@ body {
 .container {
     background-size: 100%;
     background: black;
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100vw;
     height: 100vh;
     margin: 0;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
-
 }
 
 .🍺 {
