+++ src/components/misc/LoadingScreen.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import '../../../public/sass/style.scss'
+
+const LoadingScreen = () => {
+  return (
+    <div className='container'>
+      <div className='🍺' />
+    </div>
+  )
+}
+
+export default LoadingScreen
