+++ src/views/Register/Register.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import './Register_styles.css'
+
+const Register = () => {
+  return (
+    <div>
+      <p> hola</p>
+    </div>
+  )
+}
+
+export default Register
